VANN/requirements.txt
streamlit>=1.28.0
azure-storage-blob>=12.19.0
azure-ai-documentintelligence>=1.0.0b1
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.0.0
openpyxl>=3.1.0

